Organized file structure + installed dnd
