+++ client/package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "react": "^19.1.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
