+++ client/src/index.css
@@ -1,7 +1,14 @@
 :root {
-  --dark1: #070039;
-  --dark2: #03004c;
+  --dark1: #17104b;
+  --dark2: #020128;
   --pink1: #e516b8;
-  --text-color: #000000;
+  --text-color1: #000000;
+  --text-color2: #ffffff;
+  --gray1: #9a9a9a;
   --bg-color: #ffffff;
+  --border-radius: 8px;
+}
+
+* {
+  font-family: 'Atlassian Sans Ext'
 }
